started working on accounts tab, and also refactored home component names
