+++ lib/screens/Home/Screens/Home/home.dart
@@ -6,8 +6,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:gomart/data/model/Product/product_model.dart';
-import 'package:gomart/screens/Home/Screens/HomeFragment/bloc/bloc.dart';
-import 'package:gomart/screens/Home/Screens/HomeFragment/bloc/bloc.dart';
+import 'package:gomart/screens/Home/Screens/Home/bloc/bloc.dart';
+import 'package:gomart/screens/Home/Screens/Home/bloc/bloc.dart';
 import 'package:gomart/screens/ProductDetail/product_detail.dart';
 import 'package:gomart/screens/Search/search_input.dart';
 import 'package:gomart/screens/Search/search_results.dart';
