+++ lib/screens/Home/Screens/Wallet/my_qr_screen.dart
