+++ lib/screens/Home/Screens/Wallet/scan_qr_screen.dart
