+++ lib/screens/Home/Screens/profile.dart
@@ -15,110 +15,138 @@ class ProfileFragment extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Center(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: <Widget>[
-          const Text(
-            'Test Account screen',
-          ),
-          Container(
-            padding: const EdgeInsets.fromLTRB(0, 80, 0, 0),
-            width: MediaQuery.of(context).size.width,
-            height: MediaQuery.of(context).size.height * 0.28,
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 30, vertical: 8),
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.end,
-                children: [
-                  Row(
-                    children: <Widget>[
-                      (FirebaseAuth.instance.currentUser?.photoURL != null)
-                          ? CircleAvatar(
-                              radius: 45,
-                              child: ClipOval(
-                                  child: Image.network(
-                                FirebaseAuth.instance.currentUser?.photoURL
-                                    as String,
-                                width: 120,
-                              )),
-                            )
-                          : CharacterAvatar(
-                              ch: FirebaseAuth
-                                      .instance.currentUser?.displayName ??
-                                  'Null',
+    return SafeArea(
+      child: SingleChildScrollView(
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.center,
+          mainAxisAlignment: MainAxisAlignment.start,
+          children: <Widget>[
+            Row(
+              mainAxisAlignment: MainAxisAlignment.end,
+              children: [
+                IconButton(onPressed: (){}, icon: const Icon(Icons.settings)),
+              ],
+            ),
+            Container(
+              padding: const EdgeInsets.fromLTRB(0, 4, 0, 0),
+              width: MediaQuery.of(context).size.width,
+              child: Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 30, vertical: 8),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.end,
+                  children: [
+                    Row(
+                      children: <Widget>[
+                        (FirebaseAuth.instance.currentUser?.photoURL != null)
+                            ? CircleAvatar(
+                                radius: 30,
+                                child: ClipOval(
+                                    child: Image.network(
+                                  FirebaseAuth.instance.currentUser?.photoURL
+                                      as String,
+                                  width:60,
+                                )),
+                              )
+                            : CharacterAvatar(
+                                ch: FirebaseAuth
+                                        .instance.currentUser?.displayName ??
+                                    'Null',
+                              ),
+                        const Padding(padding: EdgeInsets.all(8)),
+                        Column(
+                          mainAxisAlignment: MainAxisAlignment.start,
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: <Widget>[
+                            Text(
+                              '${FirebaseAuth.instance.currentUser?.displayName}',
+                              style: const TextStyle(
+                                  fontSize: 16, fontWeight: FontWeight.w400),
                             ),
-                      const Padding(padding: EdgeInsets.all(12)),
-                      Column(
-                        mainAxisAlignment: MainAxisAlignment.start,
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: <Widget>[
-                          Text(
-                            '${FirebaseAuth.instance.currentUser?.displayName}',
-                            style: const TextStyle(
-                                fontSize: 16, fontWeight: FontWeight.w400),
+                            const Padding(padding: EdgeInsets.all(2)),
+                            Text(
+                              '${FirebaseAuth.instance.currentUser?.phoneNumber}',
+                              style: const TextStyle(
+                                  fontSize: 14, color: Styles.colorTextDark),
+                            )
+                          ],
+                        ),
+                        Expanded(
+                          child: Row(
+                            crossAxisAlignment: CrossAxisAlignment.end,
+                            mainAxisAlignment: MainAxisAlignment.end,
+                            children:  [
+                             OutlinedButton(onPressed: (){}, child:  const Text('Edit'))
+                            ],
                           ),
-                          const Padding(padding: EdgeInsets.all(2)),
-                          Text(
-                            '${FirebaseAuth.instance.currentUser?.phoneNumber}',
-                            style: const TextStyle(
-                                fontSize: 14, color: Styles.colorTextDark),
-                          )
-                        ],
-                      )
-                    ],
-                  )
-                ],
+                        )
+                      ],
+                    )
+                  ],
+                ),
               ),
             ),
-          ),
-          const Padding(padding: EdgeInsets.all(50)),
-          ElevatedButton(
-              style: TextButton.styleFrom(
-                backgroundColor: Styles.colorPrimary,
-                padding:
-                    const EdgeInsets.symmetric(vertical: 8, horizontal: 40),
-                shape: const StadiumBorder(),
-              ),
-              onPressed: () {
-                HapticFeedback.vibrate();
-                context
-                    .read<AuthenticationBloc>()
-                    .add(AuthenticationLoggedOut(context));
-              },
-              child: const Text('Logout')),
-          (kDebugMode)
-              ? ElevatedButton(
-                  style: TextButton.styleFrom(
-                    backgroundColor: Styles.colorPrimary,
-                    padding:
-                        const EdgeInsets.symmetric(vertical: 8, horizontal: 40),
-                    shape: const StadiumBorder(),
-                  ),
-                  onPressed: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                        builder: (context) => const DebugAlgoliaSearchScreen(),
-                      ),
-                    );
-                  },
-                  child: const Text('Debug algolia page'))
-              : Container(),
-          (kDebugMode)
-              ? ElevatedButton(
-                  style: TextButton.styleFrom(
-                    backgroundColor: Styles.colorPrimary,
-                    padding:
-                        const EdgeInsets.symmetric(vertical: 8, horizontal: 40),
-                    shape: const StadiumBorder(),
-                  ),
-                  onPressed: () {
-                    context.read<DebugCubit>().seedSampleProducts();
-                  },
-                  child: const Text('Debug seed sample data'))
-              : Container()
-        ],
+            const Padding(padding: EdgeInsets.all(20)),
+            OutlinedButton(onPressed: (){}, child:  const Text('Start selling')),
+            const Padding(padding: EdgeInsets.all(20)),
+            Container(
+              color: Styles.colorWhite,
+              height: MediaQuery.of(context).size.height *0.3,
+              child:    ListView(children: const [
+                ListTile(leading: Icon(Icons.help_center), title: Text('Help Center')),
+                ListTile(leading: Icon(Icons.help_center), title: Text('App Feedback')),
+                ListTile(leading: Icon(Icons.help_center), title: Text('About Gomart')),
+                ListTile(leading: Icon(Icons.help_center), title: Text('Logout'))
+              ]),
+            ),
+            const Padding(padding: EdgeInsets.all(50)),
+            (kDebugMode && false)
+                ? ElevatedButton(
+                style: TextButton.styleFrom(
+                  backgroundColor: Styles.colorPrimary,
+                  padding:
+                      const EdgeInsets.symmetric(vertical: 8, horizontal: 40),
+                  shape: const StadiumBorder(),
+                ),
+                onPressed: () {
+                  HapticFeedback.vibrate();
+                  context
+                      .read<AuthenticationBloc>()
+                      .add(AuthenticationLoggedOut(context));
+                },
+                child: const Text('Logout')):Container(),
+            (kDebugMode && false)
+                ? ElevatedButton(
+                    style: TextButton.styleFrom(
+                      backgroundColor: Styles.colorPrimary,
+                      padding:
+                          const EdgeInsets.symmetric(vertical: 8, horizontal: 40),
+                      shape: const StadiumBorder(),
+                    ),
+                    onPressed: () {
+                      Navigator.push(
+                        context,
+                        MaterialPageRoute(
+                          builder: (context) => const DebugAlgoliaSearchScreen(),
+                        ),
+                      );
+                    },
+                    child: const Text('Debug algolia page'))
+                : Container(),
+            (kDebugMode && false)
+                ? ElevatedButton(
+                    style: TextButton.styleFrom(
+                      backgroundColor: Styles.colorPrimary,
+                      padding:
+                          const EdgeInsets.symmetric(vertical: 8, horizontal: 40),
+                      shape: const StadiumBorder(),
+                    ),
+                    onPressed: () {
+                      context.read<DebugCubit>().seedSampleProducts();
+                    },
+                    child: const Text('Debug seed sample data'))
+                : Container()
+          ],
+        ),
       ),
     );
   }
