+++ lib/screens/Home/home_screen.dart
@@ -4,18 +4,18 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_phoenix/flutter_phoenix.dart';
-import 'package:gomart/screens/Home/Screens/HomeFragment/home_Fragment.dart';
+import 'package:gomart/screens/Home/Screens/Home/home.dart';
 import 'package:gomart/screens/Home/Screens/message.dart';
 import 'package:gomart/screens/Home/Screens/saved.dart';
 import 'package:gomart/screens/Home/Screens/profile.dart';
-import 'package:gomart/screens/Home/Screens/HomeFragment/home_Fragment.dart';
+import 'package:gomart/screens/Home/Screens/Home/home.dart';
 
 import '../../data/bloc/authentication/authentication_bloc.dart';
 import '../../styles/custom_home_icons.dart';
 import '../Authentication/Register/register_screen_upload_photo.dart';
 import '../../styles/styles.dart';
 import 'Screens/profile.dart';
-import 'Screens/WalletFragment/wallet.dart';
+import 'Screens/Wallet/wallet.dart';
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({
